增加了数据字典  配置 静态奖、直推奖、动态奖代码 (未完成)

+++ src/main/java/pro/bechat/wallet/publics/RewardConfigureUtils.java
@@ -106,7 +106,7 @@ public class RewardConfigureUtils {
     /**
      * 出局制度推荐人数
      */
-    private BigDecimal outRecommendNumbers;
+    private Integer outRecommendNumbers;
 
     /**
      * 出局制本金限制
@@ -212,11 +212,11 @@ public class RewardConfigureUtils {
         this.dayLimit = dayLimit;
     }
 
-    public BigDecimal getOutRecommendNumbers() {
+    public Integer getOutRecommendNumbers() {
         return outRecommendNumbers;
     }
 
-    public void setOutRecommendNumbers(BigDecimal outRecommendNumbers) {
+    public void setOutRecommendNumbers(Integer outRecommendNumbers) {
         this.outRecommendNumbers = outRecommendNumbers;
     }
